api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
aiofiles==24.1.0
GitPython==3.1.43
PyMySQL==1.1.1
redis==5.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
jinja2==3.1.4
httpx==0.27.2
psutil==6.0.0

